refactor wcl routes + add remaining routes
the usage of tables for graph data in the frontend needs to be cleaned
up. the v1 api does NOT support graphs so this is currently a breaking
change but we can potentially work around it if it is endemic. it is a
very easy fix (just change the route from `tables` to `graph`) but not
sure how widespread it is yet
