+++ .env.local.example
@@ -1,21 +1,16 @@
-#PORT=3001
-#MYSQL_HOST=127.0.0.1
-#MYSQL_PORT=3306
-#MYSQL_DIALECT=mariadb
-#MYSQL_USER=root
-#MYSQL_ROOT_PASSWORD=my-secret-pw
-#MYSQL_DATABASE=wowanalyzer
+GITHUB_CLIENT_ID=
+GITHUB_CLIENT_SECRET=
+GITHUB_CALLBACK_URL=http://localhost:3001/login/github/callback
+LOGIN_REDIRECT_LOCATION=http://localhost:3001/
 
 PATREON_CLIENT_ID=
 PATREON_CLIENT_SECRET=
 PATREON_CALLBACK_URL=http://localhost:3001/login/patreon/callback
 
-GITHUB_CLIENT_ID=
-GITHUB_CLIENT_SECRET=
-GITHUB_CALLBACK_URL=http://localhost:3001/login/github/callback
-
-LOGIN_REDIRECT_LOCATION=http://localhost:3000/premium
-
-WCL_API_KEY=
 BATTLE_NET_API_CLIENT_ID=
 BATTLE_NET_API_CLIENT_SECRET=
+
+WCL_CLIENT_ID=
+WCL_CLIENT_SECRET=
+
+WCL_V1_API_KEY=
